fix(lint): clear the cppcheck and ruff gates CI runs
The previous commit introduced two findings that each fail `make verify`
on their own — and `make DRIFTY_STRICT=1 verify` is exactly what the
Windows CI job runs.

physics_tests.c:338 assigned a float computation to the double
TelemetryRow.timeS, flagged by cppcheck as truncLongCastAssignment. The
sibling fixture in the same file already uses (double)i * 0.1; this now
matches it. Values are identical (0.1 * n).

run_suite.py:209 exceeded ruff's 100-column limit (E501) on the new
per-lap PASS line. Split the f-string.

Verified: `make verify` passes — format, lint-py, lint (cppcheck),
analyze, test-physics, and regression (46/46 baseline comparisons,
every checksum identical).

+++ tests/scenarios/physics_tests.c
@@ -335,7 +335,7 @@ static void scenario_run_report(void)
         TelemetryRow laps[27];
         memset(laps, 0, sizeof(laps));
         for (int i = 0; i < 27; i++) {
-            laps[i].timeS = (float)(i + 1) * 0.1f;
+            laps[i].timeS = (double)(i + 1) * 0.1;
             laps[i].speedMps = 20.0f;
             laps[i].onTrack = 1;
         }
